Is this a better link to ROS documentation?
When I was reading this I was looking for info on what is ROS and an aswer to when do a product team need to do a ROS assessment. The old link did not answer that, this one does, or rateher the page links to other pages that does, including the page the old link pointed to. Is this an improvenment, perhaps?