Allow user to specify automatic redirect
This feature adds a button to the main page of the generic doc which
allows the user to configure a custom organization documentation to
redirect to when visiting the generic doc.

When this is configured, the user will be presented with a dialog
explaining that the user will be redirected, and the user can cancel the
redirect.

This is to prevent misconfiguration and permanent bad redirects.